third_party/blink/web_tests/fast/css/font-shorthand-mix-inherit-expected.txt

Test for WebKit bug 20181 : font shorthand with inherit keyword incorrectly parsed and rendered

PASS test.style.getPropertyValue('font') is '12pt / 14pt sans-serif'
PASS test.style.getPropertyValue('font') is 'inherit'
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is ''
PASS test.style.getPropertyValue('font') is '12pt / 14pt "bold inherit"'
PASS test.style.getPropertyValue('font') is 'italic 12pt / 14pt "bold inherit"'
PASS test.style.getPropertyValue('font') is 'italic 12pt / 14pt "bold arial inherit"'
PASS successfullyParsed is true

TEST COMPLETE
